feat(SMP-322): align budget aggregates server + client (topic 03, B1 B5 B15)
## Problem
- B1 — `TripsService.get_trip_home` posait `totalExpenses = 0.0` en
  dur. Le widget Dépenses sur la home n'a jamais reflété la réalité,
  même quand des BudgetItems étaient confirmés.
- B5 — le bloc wizard accumulait le breakdown SSE en `int` via
  `raw.toInt()`, perdant les décimales (jusqu'à -2,50 € sur un
  breakdown de 5 catégories à 0,50 € chacune).
- B15 — le rollback optimistic en cas de Failure réémettait `loaded`
  (l'état d'avant l'optimistic, pris depuis la mémoire bloc), mais
  ne capturait pas explicitement le `budgetSummary` antérieur, ce
  qui rendait le restore fragile si l'optimistic avait écrit dans
  d'autres zones du state entre-temps.

## Solution
- B1 : `get_trip_home` consomme `BudgetItemService.get_budget_summary
  (db, trip)["total_spent"]`. La home et le panel budget convergent
  désormais sur une seule formule.
- B5 : `TripPlan.budgetEur` typé `double`, somme du breakdown en
  `double`, fallback `total_max`/`total_min` également en `double`.
  Le cast vers `int` n'est plus fait qu'au moment du rendu.
  `step_review_view._perPersonLabel` typé `double`.
- B15 : capture explicite de `preOptimisticSummary` avant le mutate,
  restore strict via `copyWith(budgetSummary: preOptimisticSummary,
  clearBudgetSummary: preOptimisticSummary == null, operationError:
  error)` sur Failure.

> B11 (currency mixte) reste pour le topic 04b — il dépend d'un
> `currency_service` central que ce topic ne livre pas.

## Test plan
- [x] `tests/services` : 1107 passed (suite complète API).
- [x] `test/blocs/plan_trip_bloc_test.dart` : nouveau cas B5 (5 catégories
      × 0,5 € → budgetEur ≈ 2,5).
- [x] `flutter analyze` : No issues.
- [ ] Test manuel : trip avec 3 dépenses confirmées → home affiche
      le total, plus 0 €.
- [ ] Test manuel : POST /budget-items réseau ko → la card budget
      retrouve l'état pré-optimistic + bandeau erreur, pas un
      total fantôme.

+++ bagtrip/lib/plan_trip/bloc/plan_trip_bloc.dart
@@ -1003,8 +1003,11 @@ class PlanTripBloc extends Bloc<PlanTripEvent, PlanTripState> {
         .map((b) => (b['reason'] ?? '') as String)
         .toList();
 
-    // Budget total — sum breakdown categories for consistency with chart
-    int budgetEur = 0;
+    // Budget total — sum breakdown categories for consistency with chart.
+    // Topic 03 (B5) — accumulate as `double` so we don't drop the decimals
+    // (we used to lose up to ~2.50 € on a 5-category plan via
+    // `raw.toInt()`). Cast happens at render time only.
+    double budgetEur = 0;
     for (final key in [
       'flights',
       'accommodation',
@@ -1015,15 +1018,15 @@ class PlanTripBloc extends Bloc<PlanTripEvent, PlanTripState> {
       final value = budget[key];
       if (value is Map) {
         final raw = value['amount'];
-        if (raw is num) budgetEur += raw.toInt();
+        if (raw is num) budgetEur += raw.toDouble();
       } else if (value is num) {
-        budgetEur += value.toInt();
+        budgetEur += value.toDouble();
       }
     }
     if (budgetEur == 0) {
       // Fallback to LLM totals if no breakdown entries
-      final totalMax = (budget['total_max'] as num?)?.toInt() ?? 0;
-      final totalMin = (budget['total_min'] as num?)?.toInt() ?? 0;
+      final totalMax = (budget['total_max'] as num?)?.toDouble() ?? 0;
+      final totalMin = (budget['total_min'] as num?)?.toDouble() ?? 0;
       budgetEur = totalMax > 0 ? totalMax : totalMin;
     }
     if (budgetEur == 0) {
@@ -1033,7 +1036,7 @@ class PlanTripBloc extends Bloc<PlanTripEvent, PlanTripState> {
         final cost = a['estimated_cost'];
         if (cost is num) fallbackTotal += cost;
       }
-      budgetEur = fallbackTotal.toInt();
+      budgetEur = fallbackTotal;
     }
 
     return TripPlan(
